feat(.service): 会话模块接口

+++ internal/service/analysis_service.py
@@ -156,7 +156,7 @@ class AnalysisService(BaseService):
 
             # 4.确保上一时期的数据不为0，避免/0错误
             if previous_value != 0:
-                pop[field] = (current_value - previous_value) / previous_value
+                pop[field] = float((current_value - previous_value) / previous_value)
             else:
                 # 5.如果上一期数据为0，则直接设置环比为0
                 pop[field] = 0
@@ -208,7 +208,7 @@ class AnalysisService(BaseService):
             if conversation_count != 0:
                 avg_of_conversation_messages_trend_y_axis = total_messages_trend_y_axis / conversation_count
             avg_of_conversation_messages_trend["x_axis"].append(int(trend_start_at.timestamp()))
-            avg_of_conversation_messages_trend["y_axis"].append(avg_of_conversation_messages_trend_y_axis)
+            avg_of_conversation_messages_trend["y_axis"].append(float(avg_of_conversation_messages_trend_y_axis))
 
             # 8.计算费用消耗趋势图
             cost_consumption_trend_y_axis = sum(
@@ -216,7 +216,7 @@ class AnalysisService(BaseService):
                 if trend_start_at <= message.created_at < trend_end_at
             )
             cost_consumption_trend["x_axis"].append(int(trend_start_at.timestamp()))
-            cost_consumption_trend["y_axis"].append(cost_consumption_trend_y_axis)
+            cost_consumption_trend["y_axis"].append(float(cost_consumption_trend_y_axis))
 
         # 9.返回数据
         return {
